broke zssp into files
